requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.58.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.42.1
nest-asyncio==1.6.0
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pillow==11.2.1
plotly==6.1.2
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
retrying==1.3.4
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.0.6
zipp==3.23.0
